Rename the main file
